Added wiki pages to the new API.

+++ peach/api_views.py
@@ -16,8 +16,8 @@ from django.http import HttpResponse, Http404, HttpResponseServerError, HttpResp
 
 from dynamicresponse.json_response import JsonResponse
 
-from api_forms import NamespaceForm, CreateNamespaceForm
-from models import Namespace
+from api_forms import NamespaceForm, CreateNamespaceForm, CreateWikiPageForm, WikiPageForm
+from models import Namespace, WikiPage
 
 def namespaces(request):
 	if request.method == 'POST':
@@ -28,12 +28,54 @@ def namespaces(request):
 			return JsonResponse(namespace)
 		else:
 			return HttpResponse(status=400)
+	if request.method != 'GET': return HttpResponseServerError()
 	return JsonResponse(Namespace.objects.all())
 
 def namespace(request, name):
 	namespace = get_object_or_404(Namespace, name=name)
+	if not namespace.can_read(request.user): return HttpResponse(status=403)
 	if request.method == 'DELETE':
 		if not namespace.can_delete(request.user): return HttpResponse(status=403)
 		namespace.delete()
 		return HttpResponse()
+	if request.method != 'GET': return HttpResponseServerError()
 	return JsonResponse(namespace)
+
+def pages(request, name):
+	namespace = get_object_or_404(Namespace, name=name)
+	if not namespace.can_read(request.user): return HttpResponse(status=403)
+
+	if request.method == 'POST':
+		if not namespace.can_update(request.user): return HttpResponse(status=403)
+		create_wiki_page_form = CreateWikiPageForm(request.POST, instance=WikiPage(namespace=namespace))
+		if create_wiki_page_form.is_valid():
+			wiki_page = create_wiki_page_form.save()
+			if wiki_page == None: return HttpResponse(status=400)
+			return JsonResponse(wiki_page)
+		else:
+			return HttpResponse(status=400)
+	
+	if request.method != 'GET': return HttpResponseServerError()
+	return JsonResponse(namespace.pages.all())
+
+def page(request, id):
+	page = get_object_or_404(WikiPage, pk=id)
+	if not page.namespace.can_read(request.user): return HttpResponse(status=403)
+
+	if request.method == 'PUT' or request.method == 'POST':
+		if not page.namespace.can_update(request.user): HttpResponse(status=403)
+		wiki_page_form = WikiPageForm(request.POST, instance=page)
+		if wiki_page_form.is_valid():
+			page = wiki_page_form.save()
+			return JsonResponse(page)
+		else:
+			return HttpResponse(status=400)
+
+	if request.method == 'DELETE':
+		if not page.namespace.can_update(request.user): HttpResponse(status=403)
+		page.delete()
+		return HttpResponse(status=200)
+	
+	if request.method != 'GET': return HttpResponseServerError()
+	return JsonResponse(page)
+
